* commit '40433d51':
  DAQ-57 Remove "module load global/http_proxy" which is no longer needed
  JavaToUnix: change Unix.getFreeMB to return a long instead of int
  JavaToUnix: Extract JDK include location to variable in Makefile
  Update bleclipse for Eclipse Mars release to match buckminster.

+++ uk.ac.gda.dls/src/gda/util/Unix.java
@@ -184,7 +184,7 @@ public class Unix {
 	 *            to free memory blocks
 	 * @return Status code
 	 */
-	public native int doGetFreeMB(String path);
+	public native long doGetFreeMB(String path);
 
 	/**
 	 * JNI implementation specific.
@@ -442,7 +442,7 @@ public class Unix {
 	 *            Directory path of a file on the required disk.
 	 * @return Free space on the disk in MBytes.
 	 */
-	public int getFreeMB(String path) {
+	public long getFreeMB(String path) {
 		return (doGetFreeMB(path));
 	}
